feat: db_interactions.py added back
created database interface again, now working

Type: refactor
